Refactor matches_count, include Apply_color module
Included Apply_color module to Game class and moved its file in lib/
directory. Used colorize_pegs method on CLI outputs.
Refactored matches_count to avoid over counting pegs of the same color.
The new logic creates a copy of guess and each iteration if a match is
found the corresponding peg is removed.

[Bonus explanation] A guess_copy is necessary because @guess is an
instance variable available to all methods and even if it is regenerated
each turn by the play loop, it needs to be passed to the game_over?
method and therefore cannot be changed by deleting pegs from it.