Refactor code, separate classes
Refactored main.rb to handle new classes. Corrected minor bug in
play_again? by switching while loop to if else statement.
The variable name is now stored once at the begging of the game and
passed to the various methods and classes.
Kept common game interface and initialization in the superclass
GameLogic.
Split the logic of the game between Guesser and Creator classes.